@@ -10,44 +10,47 @@
   },
   "dependencies": {
     "accounting": "^0.4.1",
-    "babel-core": "^6.5.2",
-    "babel-loader": "6.2.3",
-    "babel-polyfill": "6.5.0",
-    "babel-preset-es2015": "6.5.0",
-    "babel-preset-react": "6.5.0",
-    "babel-preset-stage-0": "6.5.0",
-    "babel-relay-plugin": "0.7.3",
-    "classnames": "^2.2.3",
+    "babel-core": "^6.8.0",
+    "babel-loader": "^6.2.4",
+    "babel-polyfill": "^6.8.0",
+    "babel-preset-es2015": "^6.6.0",
+    "babel-preset-react": "^6.5.0",
+    "babel-preset-stage-0": "^6.5.0",
+    "babel-relay-plugin": "^0.8.1",
+    "classnames": "^2.2.5",
     "compression": "^1.6.1",
     "debug": "^2.2.0",
-    "express": "4.13.4",
-    "express-graphql": "0.4.9",
+    "express": "^4.13.4",
+    "express-graphql": "^0.5.1",
     "faker": "^3.1.0",
-    "graphql": "0.4.17",
-    "graphql-relay": "0.3.6",
-    "lodash.times": "^4.0.2",
-    "moment": "^2.12.0",
+    "graphql": "^0.6.0",
+    "graphql-relay": "^0.4.1",
+    "lodash.times": "^4.1.0",
+    "moment": "^2.13.0",
     "morgan": "^1.7.0",
     "ms": "^0.7.1",
-    "pg": "^4.5.2",
+    "pg": "latest",
     "pg-hstore": "^2.3.2",
-    "react": "^15.0.1",
-    "react-dom": "^15.0.1",
-    "react-relay": "0.7.3",
-    "react-router": "^2.3.0",
-    "react-router-relay": "^0.9.5",
-    "sequelize": "^3.20.0",
-    "webpack": "1.12.13",
-    "webpack-dev-server": "1.14.1"
+    "react": "^15.0.2",
+    "react-dom": "^15.0.2",
+    "react-relay": "^0.8.1",
+    "react-router": "^2.4.0",
+    "react-router-relay": "^0.13.2",
+    "redbox-react": "^1.2.4",
+    "sequelize": "^3.23.0",
+    "webpack": "^1.13.0",
+    "webpack-dev-server": "^1.14.1"
   },
   "devDependencies": {
-    "babel-cli": "6.5.1",
-    "babel-register": "^6.7.2",
-    "eslint": "^2.5.3",
-    "eslint-config-airbnb": "^6.2.0",
-    "eslint-plugin-react": "^4.2.3",
+    "babel-cli": "^6.8.0",
+    "babel-register": "^6.8.0",
+    "eslint": "^2.9.0",
+    "eslint-config-airbnb": "^9.0.1",
+    "eslint-plugin-import": "^1.8.0",
+    "eslint-plugin-jsx-a11y": "^1.2.0",
+    "eslint-plugin-react": "^5.1.1",
     "proxy-middleware": "^0.15.0",
-    "react-hot-loader": "^1.3.0",
+    "react-hot-loader": "^3.0.0-beta.1",
     "redbox-react": "^1.2.2",
     "webpack-dev-middleware": "^1.6.1",
     "webpack-hot-middleware": "^2.10.0"
