@@ -1,41 +1,29 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
+import { render } from 'react-dom'
+import { AppContainer } from 'react-hot-loader'
 
-// import { browserHistory } from 'react-router'
-// import routes from './routes'
-const rootEl = document.getElementById('root')
 
-const render = () => {
-  const Root = require('./root').default
-  ReactDOM.render(
-    <Root />,
-    rootEl
+import Root from './Root'
+const rootElement = document.getElementById('root')
+
+render(
+  <AppContainer>
+    <Root />
+  </AppContainer>,
+  rootElement
+)
+
+if (module.hot) {
+  module.hot.accept(
+    './Root',
+    () => {
+      const NextRoot = require('./Root').default
+      render(
+        <AppContainer>
+          <NextRoot />
+        </AppContainer>,
+        rootElement
+      )
+    }
   )
 }
-
-let finalRender = render
-if (module.hot) {
-  const renderApp = () => {
-    // ReactDOM.unmountComponentAtNode(rootEl)
-    render()
-  }
-  const renderError = (error) => {
-    const RedBox = require('redbox-react')
-    ReactDOM.render(
-      <RedBox error={error} />,
-      rootEl
-    )
-  }
-  finalRender = () => {
-    try {
-      renderApp()
-    } catch (error) {
-      renderError(error)
-    }
-  }
-  module.hot.accept('./root', () => {
-    setTimeout(render)
-  })
-}
-
-finalRender()
